refactor: 优化代码结构和命名，更新Kubernetes相关逻辑
- 移除不必要的导入和重复的常量定义，简化代码。
- 更新HostInstallController以支持sshPassword参数。
- 修改NodeExportUtils中的K8S为Kubernetes，确保一致性。
- 删除K8sDeploymentYamlHandler和K8sServiceStartHandler类，整合相关逻辑。
- 更新UI组件中的K8S为Kubernetes，增强可读性和一致性。
- 添加新的Kubernetes配置文件以支持SparkOperator。

这些更改旨在提高代码的可维护性和一致性，同时确保Kubernetes相关功能的正常运行。

+++ datasophon-service/src/main/java/com/datasophon/api/service/host/dto/QueryHostListPageDTO.java
@@ -70,7 +70,7 @@ public class QueryHostListPageDTO {
 
     private String nodeLabel;
 
-    private Integer serviceRoleNum;
+    private Long serviceRoleNum;
 
     /**
      * 主机的hosts文件内容
